move permissions.js
